extract cleanup data table
